+++ lite_web/lite/exceptions.py
